clean ups
* omit raw types where possibly
* omit redundant use of Optionals
* omit auto (un-)boxing where the result is clear
* switch to standard library methods where possible

+++ api/src/main/java/net/automatalib/graphs/helpers/SimpleNodeIDs.java
@@ -47,7 +47,7 @@ public class SimpleNodeIDs<N> implements NodeIDs<N> {
             throw new IllegalArgumentException();
         }
 
-        return id.intValue();
+        return id;
     }
 
     @Override
